solved import bugs, created vulkan instance

+++ include/texture.hpp
@@ -5,8 +5,8 @@
 
 typedef struct Texture {
     char* data;
-    size_t height;
-    size_t width;
+    int height;
+    int width;
 
     Texture();
 } Texture;
